Move parameter mapping, private petab helpers
* Move petab simulation code from `.importers` to `.sim`
* Separate sundials and non-sundials related code so the jax-related modules don't import the cpp extension
* Make helpers private

Closes #3041.
