+++ python/sdist/amici/sim/__init__.py
@@ -1,1 +1,3 @@
 """Functionality for simulating AMICI models."""
+
+from __future__ import annotations
