+++ python/sdist/amici/sim/sundials/petab/v1/_petab_problem.py
@@ -9,10 +9,12 @@ import petab.v1 as petab
 from petab.v1.C import PREEQUILIBRATION_CONDITION_ID, SIMULATION_CONDITION_ID
 
 import amici
-from amici.importers.petab.v1.parameter_mapping import create_parameter_mapping
 from amici.sim.sundials.petab.v1._conditions import (
     create_edatas,
     fill_in_parameters,
+)
+from amici.sim.sundials.petab.v1._parameter_mapping import (
+    create_parameter_mapping,
 )
