+++ python/sdist/amici/sim/sundials/petab/v1/_simulator.py
@@ -18,7 +18,7 @@ from collections.abc import Callable
 import pandas as pd
 import petab.v1 as petab
 
-from amici.importers.petab.v1.petab_import import import_petab_problem
+from amici.importers.petab.v1._petab_import import import_petab_problem
 from amici.sim.sundials import AmiciModel, SensitivityMethod
 
 from ._simulations import RDATAS, rdatas_to_measurement_df, simulate_petab
