+++ python/tests/adapters/test_fiddy.py
@@ -2,15 +2,13 @@
 
 from pathlib import Path
 
-import amici
-import amici.importers.petab.v1.petab_import
-import amici.sim.sundials.petab.v1._simulations
 import numpy as np
 import pytest
 from amici.adapters.fiddy import (
     run_simulation_to_cached_functions,
     simulate_petab_to_cached_functions,
 )
+from amici.importers.petab.v1 import import_petab_problem
 from amici.sim.sundials import SensitivityOrder, SteadyStateSensitivityMode
 from fiddy import MethodId, Type, get_derivative
 from fiddy.derivative_check import NumpyIsCloseDerivativeCheck
@@ -41,9 +39,7 @@ def lotka_volterra() -> tuple[v1.Problem, np.ndarray]:
 def test_run_amici_simulation_to_functions(problem_generator):
     petab_problem, point = problem_generator()
     timepoints = sorted(set(petab_problem.measurement_df.time))
-    amici_model = amici.importers.petab.v1.petab_import.import_petab_problem(
-        petab_problem
-    )
+    amici_model = import_petab_problem(petab_problem)
     amici_model.set_timepoints(timepoints)
     amici_solver = amici_model.create_solver()
 
@@ -108,9 +104,7 @@ def test_run_amici_simulation_to_functions(problem_generator):
 @pytest.mark.parametrize("scaled_parameters", (False, True))
 def test_simulate_petab_to_functions(problem_generator, scaled_parameters):
     petab_problem, point = problem_generator()
-    amici_model = amici.importers.petab.v1.petab_import.import_petab_problem(
-        petab_problem
-    )
+    amici_model = import_petab_problem(petab_problem)
     amici_solver = amici_model.create_solver()
 
     if amici_model.get_name() == "simple":
